Auto stash before merge of "main" and "origin/main"

+++ src/main/java/car/Acceleration.java
@@ -1,6 +1,6 @@
 package car;
 
 public interface Acceleration {
-    void acceleration(double grip, int distance, double reqSpeed);
+    void acceleration(double grip, double distance, double reqSpeed);
     double braking(double grip, double reqSpeed);
 }
